Add Operation1 as a base class for single-argument operations
