+++ h2/src/main/org/h2/expression/Operation1.java
@@ -1,0 +1,69 @@
+/*
+ * Copyright 2004-2020 H2 Group. Multiple-Licensed under the MPL 2.0,
+ * and the EPL 1.0 (https://h2database.com/html/license.html).
+ * Initial Developer: H2 Group
+ */
+package org.h2.expression;
+
+import org.h2.engine.Session;
+import org.h2.table.ColumnResolver;
+import org.h2.table.TableFilter;
+import org.h2.value.TypeInfo;
+
+/**
+ * Operation with one argument.
+ */
+public abstract class Operation1 extends Expression {
+
+    protected Expression arg;
+
+    protected TypeInfo type;
+
+    protected Operation1(Expression arg) {
+        this.arg = arg;
+    }
+
+    @Override
+    public TypeInfo getType() {
+        return type;
+    }
+
+    @Override
+    public void mapColumns(ColumnResolver resolver, int level, int state) {
+        arg.mapColumns(resolver, level, state);
+    }
+
+    @Override
+    public void setEvaluatable(TableFilter tableFilter, boolean value) {
+        arg.setEvaluatable(tableFilter, value);
+    }
+
+    @Override
+    public void updateAggregate(Session session, int stage) {
+        arg.updateAggregate(session, stage);
+    }
+
+    @Override
+    public boolean isEverything(ExpressionVisitor visitor) {
+        return arg.isEverything(visitor);
+    }
+
+    @Override
+    public int getCost() {
+        return arg.getCost() + 1;
+    }
+
+    @Override
+    public int getSubexpressionCount() {
+        return 1;
+    }
+
+    @Override
+    public Expression getSubexpression(int index) {
+        if (index == 0) {
+            return arg;
+        }
+        throw new IndexOutOfBoundsException();
+    }
+
+}
