+++ h2/src/main/org/h2/expression/UnaryOperation.java
@@ -6,8 +6,6 @@
 package org.h2.expression;
 
 import org.h2.engine.Session;
-import org.h2.table.ColumnResolver;
-import org.h2.table.TableFilter;
 import org.h2.value.TypeInfo;
 import org.h2.value.Value;
 import org.h2.value.ValueNull;
@@ -15,13 +13,10 @@ import org.h2.value.ValueNull;
 /**
  * Unary operation. Only negation operation is currently supported.
  */
-public class UnaryOperation extends Expression {
-
-    private Expression arg;
-    private TypeInfo type;
+public class UnaryOperation extends Operation1 {
 
     public UnaryOperation(Expression arg) {
-        this.arg = arg;
+        super(arg);
     }
 
     @Override
@@ -39,11 +34,6 @@ public class UnaryOperation extends Expression {
     }
 
     @Override
-    public void mapColumns(ColumnResolver resolver, int level, int state) {
-        arg.mapColumns(resolver, level, state);
-    }
-
-    @Override
     public Expression optimize(Session session) {
         arg = arg.optimize(session);
         type = arg.getType();
@@ -56,44 +46,6 @@ public class UnaryOperation extends Expression {
             return ValueExpression.get(getValue(session));
         }
         return this;
-    }
-
-    @Override
-    public void setEvaluatable(TableFilter tableFilter, boolean b) {
-        arg.setEvaluatable(tableFilter, b);
-    }
-
-    @Override
-    public TypeInfo getType() {
-        return type;
-    }
-
-    @Override
-    public void updateAggregate(Session session, int stage) {
-        arg.updateAggregate(session, stage);
-    }
-
-    @Override
-    public boolean isEverything(ExpressionVisitor visitor) {
-        return arg.isEverything(visitor);
-    }
-
-    @Override
-    public int getCost() {
-        return arg.getCost() + 1;
-    }
-
-    @Override
-    public int getSubexpressionCount() {
-        return 1;
-    }
-
-    @Override
-    public Expression getSubexpression(int index) {
-        if (index == 0) {
-            return arg;
-        }
-        throw new IndexOutOfBoundsException();
     }
 
 }
